Homeworks Topic 12, 13, 14 (hw and pt).

+++ src/com/softserve/edu05/apptest/Calc.java
@@ -7,7 +7,7 @@ public class Calc {
         //return arg0 + arg1;
     }
 
-    public double div(int arg0, int arg1) {
+    public double div(double arg0, double arg1) {
     //public double div(double arg0, double arg1) {
         return arg0 / arg1;
     }
